Fix parser for root (only) dataset names - investigate (#619)
* contrib/test-splitHostDataSet.sh: add test cases for root dataset of a pool [#585]


* lib/ZnapZend/ZFS.pm: destroySnapshots(): sanity-check for undef $dataSet [#585]


* t/znapzend-lib-splitter.t: add a pure-perl test for splitHostDataSet() and splitDataSetSnapshot() [#585]


* lib/ZnapZend/ZFS.pm: splitDataSetSnapshot(): refine to check if argument is actually a "dataset@snapname", return an undef snapname in the array if not [#585]


* test.sh: call t/znapzend-lib-splitter.t (but ignore faults for now) [#585]


---------

+++ lib/ZnapZend/ZFS.pm
@@ -54,7 +54,12 @@ my $splitHostDataSet = sub {
 };
 
 my $splitDataSetSnapshot = sub {
-    return ($_[0] =~ /^([^\@]+)\@([^\@]+)$/);
+    my $count = ($_[0] =~ tr/@//);
+    if ($count > 0) {
+        return ($_[0] =~ /^([^\@]+)\@([^\@]+)$/);
+    } else {
+        return ($_[0], undef);
+    }
 };
 
 my $shellQuote = sub {
@@ -368,11 +373,16 @@ sub destroySnapshots {
         for my $task (@toDestroy){
             my ($remote, $dataSetPathAndSnap) = $splitHostDataSet->($task);
             my ($dataSet, $snapshot) = $splitDataSetSnapshot->($dataSetPathAndSnap);
-            my @ssh = $self->$buildRemote($remote, [@{$self->priv}, qw(zfs destroy), @recursive, "$dataSet\@$snapshot"]);
+            if (defined ($dataSet)) {
+                my @ssh = $self->$buildRemote($remote, [@{$self->priv}, qw(zfs destroy), @recursive, "$dataSet\@$snapshot"]);
 
-            print STDERR '# ' . (($self->noaction || $self->nodestroy) ? "WOULD # " : "") . join(' ', @ssh) . "\n" if $self->debug;
-            system(@ssh) and $destroyError .= "ERROR: cannot destroy snapshot $dataSet\@$snapshot\n"
-                if !($self->noaction || $self->nodestroy);
+                print STDERR '# ' . (($self->noaction || $self->nodestroy) ? "WOULD # " : "") . join(' ', @ssh) . "\n" if $self->debug;
+                system(@ssh) and $destroyError .= "ERROR: cannot destroy snapshot $dataSet\@$snapshot\n"
+                    if !($self->noaction || $self->nodestroy);
+            } else {
+                print STDERR "[D] task='$task' => remote='$remote' dataSetPathAndSnap='$dataSetPathAndSnap' => dataSet='$dataSet' snapshot='$snapshot'\n";
+                Mojo::Exception->throw("ERROR: oracleMode destroy: failed to parse task='$task', got undefined dataSet and/or snapshot");
+            }
         }
         #remove trailing \n
         chomp $destroyError;
